todo_woo/requirements.txt
asgiref==3.5.2
Django==4.1.1
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.4
djangorestframework==3.13.1
